Grouping functionality - summing losses and exposures, still not generating the matrice

+++ RR-Inuring/src/main/java/com/scor/rr/JsonFormat/Plts.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Plts {
 
-    private int pltId;
+    private long pltId;
     private String pltFilePath;
     private String pltFileName;
     private String ccy;
